警告対応

+++ cppshogi/common.hpp
@@ -193,7 +193,7 @@ FORCE_INLINE int msb(const u64 b) {
 #if defined(HAVE_SSE42)
 #include <nmmintrin.h>
 inline int count1s(u64 x) {
-    return _mm_popcnt_u64(x);
+    return static_cast<int>(_mm_popcnt_u64(x));
 }
 #else
 inline int count1s(u64 x) //任意の値の1のビットの数を数える。( x is not a const value.)
